document strcpy alternatives, change refcounts to long

+++ src/sentry_value.c
@@ -40,7 +40,7 @@
 
 typedef struct {
     void *payload;
-    int refcount;
+    long refcount;
     char type;
 } thing_t;
